+++ internal/framework/provider/provider.go
@@ -7,6 +7,7 @@ import (
 	"math"
 	"regexp"
 	"strconv"
+	"strings"
 	"testing"
 
 	cfv1 "github.com/cloudflare/cloudflare-go"
@@ -20,6 +21,8 @@ import (
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/dlp_datasets"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/email_routing_address"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/email_routing_rule"
+	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/gateway_app_types"
+	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/gateway_categories"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/hyperdrive_config"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/list_item"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/origin_ca_certificate"
@@ -28,7 +31,8 @@ import (
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/rulesets"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/turnstile"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/user"
-	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/workers_for_platforms"
+	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/workers_for_platforms_dispatch_namespace"
+	"github.com/cloudflare/terraform-provider-cloudflare/internal/framework/service/workers_for_platforms_dispatch_namespace_deprecated"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/sdkv2provider"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/utils"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
@@ -194,8 +198,12 @@ func (p *CloudflareProvider) Configure(ctx context.Context, req provider.Configu
 	} else {
 		basePath = utils.GetDefaultFromEnv(consts.APIBasePathEnvVarKey, consts.APIBasePathDefault)
 	}
+
 	baseURL := cfv1.BaseURL(fmt.Sprintf("https://%s%s", baseHostname, basePath))
-	cfv2Options = append(cfv2Options, option.WithBaseURL(fmt.Sprintf("https://%s%s", baseHostname, basePath)))
+
+	// Ensure there is a trailing slash for client.V2 basePath
+	basePathV2 := strings.TrimSuffix(basePath, "/") + "/"
+	cfv2Options = append(cfv2Options, option.WithBaseURL(fmt.Sprintf("https://%s%s", baseHostname, basePathV2)))
 
 	if !data.RPS.IsNull() {
 		rps = int64(data.RPS.ValueInt64())
@@ -362,7 +370,8 @@ func (p *CloudflareProvider) Resources(ctx context.Context) []func() resource.Re
 		rulesets.NewResource,
 		turnstile.NewResource,
 		access_mutual_tls_hostname_settings.NewResource,
-		workers_for_platforms.NewResource,
+		workers_for_platforms_dispatch_namespace_deprecated.NewResource,
+		workers_for_platforms_dispatch_namespace.NewResource,
 	}
 }
 
@@ -372,6 +381,8 @@ func (p *CloudflareProvider) DataSources(ctx context.Context) []func() datasourc
 		origin_ca_certificate.NewDataSource,
 		user.NewDataSource,
 		dlp_datasets.NewDataSource,
+		gateway_categories.NewDataSource,
+		gateway_app_types.NewDataSource,
 	}
 }