+++ internal/framework/service/workers_for_platforms_dispatch_namespace_deprecated/model.go
@@ -1,8 +1,8 @@
-package workers_for_platforms
+package workers_for_platforms_dispatch_namespace_deprecated
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
 
-type WorkersForPlatformsNamespaceModel struct {
+type WorkersForPlatformsNamespaceDeprecatedModel struct {
 	AccountID types.String `tfsdk:"account_id"`
 	Name      types.String `tfsdk:"name"`
 	ID        types.String `tfsdk:"id"`