+++ internal/sdkv2provider/provider.go
@@ -21,7 +21,7 @@ import (
 const (
 	MAXIMUM_NUMBER_OF_ENTITIES_REACHED_SUMMARY = "You've attempted to add a new %[1]s to the `terraform-plugin-sdkv2` which is no longer considered suitable for use."
 	MAXIMUM_NUMBER_OF_ENTITIES_REACHED_DETAIL  = "Due the number of known internal issues with `terraform-plugin-sdkv2` (most notably handling of zero values), we are no longer recommending using it and instead, advise using `terraform-plugin-framework` exclusively. If you must use terraform-plugin-sdkv2 for this new %[1]s you should first discuss it with a maintainer to fully understand the impact and potential ramifications. Only then should you bump %[2]s to include your %[1]s."
-	MAXIMUM_ALLOWED_SDKV2_RESOURCES            = 108
+	MAXIMUM_ALLOWED_SDKV2_RESOURCES            = 113
 	MAXIMUM_ALLOWED_SDKV2_DATASOURCES          = 19
 )
 
@@ -287,6 +287,11 @@ func New(version string) func() *schema.Provider {
 				"cloudflare_worker_route":                                    resourceCloudflareWorkerRoute(),
 				"cloudflare_worker_script":                                   resourceCloudflareWorkerScript(),
 				"cloudflare_worker_secret":                                   resourceCloudflareWorkerSecret(),
+				"cloudflare_workers_cron_trigger":                            resourceCloudflareWorkersCronTrigger(),
+				"cloudflare_workers_domain":                                  resourceCloudflareWorkersDomain(),
+				"cloudflare_workers_route":                                   resourceCloudflareWorkersRoute(),
+				"cloudflare_workers_script":                                  resourceCloudflareWorkersScript(),
+				"cloudflare_workers_secret":                                  resourceCloudflareWorkersSecret(),
 				"cloudflare_workers_kv_namespace":                            resourceCloudflareWorkersKVNamespace(),
 				"cloudflare_workers_kv":                                      resourceCloudflareWorkerKV(),
 				"cloudflare_zone_cache_reserve":                              resourceCloudflareZoneCacheReserve(),