+++ internal/sdkv2provider/schema_cloudflare_access_application.go
@@ -67,7 +67,6 @@ func resourceCloudflareAccessApplicationSchema() map[string]*schema.Schema {
 			},
 			Optional: true,
 			Description: "The policies associated with the application, in ascending order of precedence." +
-				" When omitted, the application policies are not be updated." +
 				" Warning: Do not use this field while you still have this application ID referenced as `application_id`" +
 				" in any `cloudflare_access_policy` resource, as it can result in an inconsistent state.",
 		},
@@ -559,6 +558,12 @@ func resourceCloudflareAccessApplicationSchema() map[string]*schema.Schema {
 					},
 				},
 			},
+		},
+		"skip_app_launcher_login_page": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "Option to skip the App Launcher landing page.",
 		},
 		"allow_authenticate_via_warp": {
 			Type:        schema.TypeBool,